+++ app/lib/cabinet-service.js
@@ -1,3 +1,5 @@
+'use strict';
+
 angular.module('lizard-nxt')
   .service("CabinetService", ["$q", "Restangular",
   function ($q, Restangular) {
@@ -9,7 +11,8 @@ angular.module('lizard-nxt')
       apiLayerGroups,
       timeseriesLocationObjectResource,
       timeseriesResource,
-      flowResource;
+      flowResource,
+      events;
 
   // for the wizard demo's
   if (window.location.host === 'nens.github.io') {
@@ -19,6 +22,7 @@ angular.module('lizard-nxt')
   geocodeResource = Restangular.one('api/v1/geocode/');
   reverseGeocodeResource = Restangular.one('api/v1/reversegeocode/');
   timeseriesResource = Restangular.one('api/v1/timeseries/');
+  events = Restangular.one('api/v1/events/');
 
   /**
    * Raster resource, last stop to the server
@@ -70,6 +74,7 @@ angular.module('lizard-nxt')
 
   return {
     //eventTypes: eventTypes,
+    events: events,
     tooltips: tooltips,
     geocode: geocodeResource,
     raster: rasterResource,